# Speed up test_es_client.py and test_utils.py slow tests