--- a/connectors/tests/test_es_client.py
+++ b/connectors/tests/test_es_client.py
@@ -8,6 +8,7 @@ import base64
 import pytest
 
 from connectors.es.client import ESClient
+from elasticsearch import ConnectionError
 
 
 def test_esclient():
@@ -79,19 +80,18 @@ async def test_es_client_auth_error(mock_responses, patch_logger):
 
 
 @pytest.mark.asyncio
-async def test_es_client_no_server(mock_responses, patch_logger):
-    host = "http://nowhere.com:9200"
-    mock_responses.get(
-        host, status=503, headers={"X-Elastic-Product": "Elasticsearch"}, repeat=True
-    )
+async def test_es_client_no_server(patch_logger):
     # if we can't reach the server, we need to catch it cleanly
     config = {
         "username": "elastic",
         "password": "changeme",
-        "host": host,
+        "host": "http://nowhere.com:9200",
         "max_wait_duration": 0.1,
         "initial_backoff_duration": 0.1,
     }
     es_client = ESClient(config)
-    assert not await es_client.ping()
-    await es_client.close()
+
+    with mock.patch.object(es_client.client, "info", side_effect=ConnectionError()):
+        # Execute
+        assert not await es_client.ping()
+        await es_client.close()