--- a/connectors/tests/test_es_client.py
+++ b/connectors/tests/test_es_client.py
@@ -79,12 +79,16 @@ async def test_es_client_auth_error(mock_responses, patch_logger):
 
 
 @pytest.mark.asyncio
-async def test_es_client_no_server(patch_logger):
+async def test_es_client_no_server(mock_responses, patch_logger):
+    host = "http://nowhere.com:9200"
+    mock_responses.get(host, status=503, headers={"X-Elastic-Product": "Elasticsearch"}, repeat=True)
     # if we can't reach the server, we need to catch it cleanly
     config = {
         "username": "elastic",
         "password": "changeme",
-        "host": "http://nowhere.com:9200",
+        "host": host,
+        "max_wait_duration": 0.1,
+        "initial_backoff_duration": 0.1,
     }
     es_client = ESClient(config)
     assert not await es_client.ping()