Similar to https://github.com/elastic/connectors-python/pull/343 speeding up two slowest tests:

1. test_es_client.py

```
> bin/pytest -sv --durations=0 connectors/tests/test_es_client.py
```

Before:

```
================================================================================================ slowest durations ================================================================================================
4.17s call     connectors/tests/test_es_client.py::test_es_client_no_server
```

After:

```
================================================================================================ slowest durations ================================================================================================
0.02s call     connectors/tests/test_es_client.py::test_es_client_no_server
0.01s setup    connectors/tests/test_es_client.py::test_es_client_no_server
```

2. connectors/tests/test_utils.py

Before:

```
================================================================================================ slowest durations ================================================================================================
2.10s call     connectors/tests/test_utils.py::test_mem_queue
```

After:

```
================================================================================================ slowest durations ================================================================================================
0.20s call     connectors/tests/test_utils.py::test_mem_queue
```